restructure to improve organization
