+++ configuration-app/src/findmypi/meson.build
@@ -1,0 +1,6 @@
+install_data(
+    'findmypi.py',
+    'findmypiserver.py',
+    'findmypiclient.py',
+    install_dir: join_paths(INSTALL_DIR, 'findmypi')
+)
