+++ configuration-app/src/lib/apthelper.py
