+++ configuration-app/src/lib/hint.py
