+++ configuration-app/src/lib/meson.build
@@ -1,0 +1,6 @@
+install_data(
+    'apthelper.py',
+    'hint.py',
+    'vncwindow.py',
+    install_dir: join_paths(INSTALL_DIR, 'lib')
+)
