+++ configuration-app/src/meson.build
@@ -1,16 +1,9 @@
 subdir('scripts')
+subdir('findmypi')
+subdir('modules')
+subdir('lib')
 
 install_data(
     'main.py',
-    'remote.py',
-    'layout.py',
-    'overclock.py',
-    'vncwindow.py',
-    'hint.py',
-    'display.py',
-    'apthelper.py',
-    'findmypi.py',
-    'findmypiserver.py',
-    'findmypiclient.py',
     install_dir: INSTALL_DIR
 )
