+++ configuration-app/src/modules/meson.build
@@ -1,0 +1,7 @@
+install_data(
+    'display.py',
+    'layout.py',
+    'overclock.py',
+    'remote.py',
+    install_dir: join_paths(INSTALL_DIR, 'modules')
+)
